requirements.txt
Flask
SQLAlchemy
pymysql
PyPDF2
sentence-transformers
chromadb
Werkzeug
Pillow
torch
ftfy
regex
tqdm
pymupdf